Well now, this is an embarassing bug that I fixed. Basically, there was a problem with the sparse matrix multiplication routine. At some point, you've computed a new contribution to a matrix element. First, you add that to the value array. Now, if that was the first contribution to that matrix element, you also need to modify the hash table. The number of elements in a bucket is increased.
So you needed a way to detect if this was the first contribution
to that matrix element. Originally, we just checked if the value
of that matrix element was zero, because the value_array is
reset during the prune operation.

The problem here is probably obvious to everyone else but me.

What happens if you cancel out a matrix element perfectly? Well
then it fails, because now we're double counting a matrix element
for that bucket. The answer is probably still correct, but in the
dense matrix limit, sometimes you end up with more elements per
bucket than there is space in the bucket. This can either lead
to writing past an array's bounds, or worse overwriting a neighbor
bucket.

So I fixed this problem by addding another buffer. This buffer
is a dirty bit buffer, which has a value of 1 for matrix elements
that have been written. The cost is more memory, and one more
read/write command. We need to evaluate the performance impact,
but it shouldn't be too bad.