Per branch testing.
This is to prevent all the clashing over the use of CurrentTest.sh . Now,
if you have a test you want to run, just make a script using the name
of the git branch inside the unit test folder. For example, if your git branch
is "MyFeature", than make UnitTests/MyFeature.sh . Inside it, you export
a variable which describes which test you want to run (see
ExampleBranchScript.sh) for how this works.