Matrix Maps Feature (#80)
* A generic procedure for mapping distributed matrices.

Unfortunately, I don't have a unit test yet so we have to keep
this out of the master branch. But for now I want to use this in
NTChem.

* Mapping interface improved

* Updated with example, fixed slicing.

* An example of the matrix mapper in C++

* Maps are fully working

* Unit test for matrix maps

* A generic procedure for mapping distributed matrices.

Unfortunately, I don't have a unit test yet so we have to keep
this out of the master branch. But for now I want to use this in
NTChem.

* Mapping interface improved

* Merge with master

* An example of the matrix mapper in C++

* Maps are fully working

* Unit test for matrix maps

* Matrix map Fortran interface: subroutine->function

This is more consistent with C++

* Support for supplementary data in maps

* Actually I was wrong, we don't need supp for C++.

Overriding works fine for C++ and Python

* Mapping interface enrichment.